Rename http directory
