+++ lib/ibm/cloud/sdk/vpc/http/base_collection.rb
