+++ lib/ibm/cloud/sdk/vpc/http/base_instance.rb
